+++ pkg/label/label.go
@@ -3,6 +3,9 @@ package label
 const (
 	Cluster      = "giantswarm.io/cluster"
 	Organization = "giantswarm.io/organization"
+
+	ManagedBy = "giantswarm.io/managed-by"
+
 	// Labels, used in legacy cluster namespaces
 	LegacyCluster  = "cluster"
 	LegacyCustomer = "customer"